requirements.txt
fairseq==0.12.2
funasr==1.3.0
hydra-core==1.3.2
librosa==0.11.0
matplotlib==3.8.4
numpy==2.4.1
omegaconf==2.3.0
pandas==2.3.3
scikit_learn==1.4.2
scipy==1.17.0
seaborn==0.13.2
soundfile==0.13.1
timm==1.0.24
torch==2.9.1
